% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_tcga.R
\name{extract_tcga_clinical}
\alias{extract_tcga_clinical}
\title{Extract clinical information from tcga format}
\usage{
extract_tcga_clinical(clin)
}
\description{
Extract clinical information from tcga format
}

