% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{plot_KMCurve}
\alias{plot_KMCurve}
\title{Plot Kaplan-Meier Curve}
\usage{
plot_KMCurve(clinical, labels, annot = NULL, color = NULL, font = "Arial",
  xlab = "Follow up (weeks)", ylab = "DFS (prob.)", legend.pos = "top",
  risk.table = F, ...)
}
\arguments{
\item{clinical}{a survival object created by Surv function in survival package}

\item{labels}{a vector containing subtyping labels of patients}

\item{annot}{a character indicating the annotation showed up in the plot}

\item{color}{a vector containing colors used for different subtypes}

\item{font}{a character indicating the font used in the plot (default: "Arial")}

\item{xlab}{a character indicating the label of x-axis (default: "Follow up (weeks)")}

\item{ylab}{a character indicating the label of y-axis (default: "DFS (prob.)")}

\item{risk.table}{a logical indicating whether show risk table or not (default: FALSE)}

\item{lenged.pos}{a character indicating whether the legend is (default: "top")}
}
\value{
a ggplot2 object of the plot
}
\description{
This function plots Kaplan-Meier Curve for survival analysis.
}
\examples{
clinical <- survival::Surv(t.rfs, e.rfs)
color.cms <- c("#E69E00","#0070B0","#CA78A6", "#009C73")
plot_KMCurve(clinical, labels, "GSE39582", color.cms)
}

